FINALLY got through the monster glitches
I added a creature test class that YES checks through things. I've
updated how the player's move according to the entityArray.

According to the tests the player, when within a close proximity can
tell the monster is there.

NOW to the basic monster testing and get the thing to wander in turn.