Detect Stairs
Made changes to the Player.java file's update method and the
BasicMap.java file all over the place (constructor and at the bottom of
the file.  When the player steps on the stair tile (bottom left of the
current map), the console will say that the player has stepped on the
stairs so it is detected.