Fixed Stairs looping case
Found a glitch in last update. When game reset it carried levels that
were not yet played over form the last game-over.
Fixed that so it clears all levels before loading again.

- Fixed entity isStairs method. There are still edges in the map...I
fixed the boundary checkers in the code just because it annoyed me
enough.

- Also, played through the game myself (w/o shadows) It is impossible to
win the game (reach floor 7) by hitting all monsters in sight. You are
bound to get unlucky with your rolls and miss several times. However it
is possible to dodge as many monsters as you can (fighting only when
necessary) and reach the last floor.

Mind you, I did this w/o shadows so it might be harder with them...