Saving & Loading Slightly Fixed.....
Just did a bit of adjusting in the loading player stuff
I'll be sending some q's about it later...

In general the error:
Level 1 = 30 hp points.  not 200 hp points...I'm not sure where that's
loading from.
- Stat levels did not match what's being loaded.

I will double check the stats in the intialization stuff. (I changed the
way stats are calculated to match what we talked about in the meeting in
my last version)